Extract blob (de)serialization into `NamedCachedQueryBlobManager`
Address the review of PR #3052:

- Move all blob (de)serialization out of the `Qlever` class into a new
  `NamedCachedQueryBlobManager` (new files `src/libqlever/NamedCachedQueryBlob
  Manager.{h,cpp}`). `Qlever` now holds one as a member and keeps only two thin
  public delegators (`serializeVocabAndNamedCacheToCompressedBlob` /
  `deserializeVocabAndNamedCacheFromCompressedBlob`). The manager is a friend of
  `Qlever`. The `BlobAllocator`, the decompressed-buffer lifetime extender, and
  the (previously `Qlever`-owned) named-result-cache (de)serialization move into
  the manager, so the core `Qlever` class shrinks.
- Split the blob format into small, independently testable static utilities:
  `compressBlobAndAddTrailingSizeInfo` (renamed inverse
  `decompressBlobWithTrailingSizeInfo`), and a `writeBlobHeader` mirroring
  `skipAndVerifyBlobHeader`. Bind `index_.getImpl()` to a reference in
  serialize/deserialize.
- Use `fileToJson` in `IndexImpl::readConfiguration` (from the previous round).
- Add unit tests that exercise the compress/decompress round-trip (several
  sizes incl. empty) and the header write/verify in isolation, and move the
  blob end-to-end tests into a new `NamedCachedQueryBlobManagerTest.cpp`.
- Add a `FRIEND_TEST` that fires the `swapIndexAndViews` assertion (swap while
  the named result cache is non-empty), kept in `QleverTest.cpp`.
